Refacto blake2f
