+++ tests/src/kakarot/precompiles/test_blake2f.cairo
@@ -1,34 +1,29 @@
-// SPDX-License-Identifier: MIT
-
 %lang starknet
 
-// Starkware dependencies
 from starkware.cairo.common.alloc import alloc
 from starkware.cairo.common.cairo_builtins import HashBuiltin, BitwiseBuiltin
 from starkware.cairo.common.memcpy import memcpy
 from starkware.cairo.common.memset import memset
 
-// Local dependencies
 from utils.utils import Helpers
 from kakarot.precompiles.blake2f import PrecompileBlake2f
 
-@external
 func test_should_fail_when_input_is_not_213{
     syscall_ptr: felt*, pedersen_ptr: HashBuiltin*, range_check_ptr, bitwise_ptr: BitwiseBuiltin*
-}() -> (output_len: felt, output: felt*) {
+}(output_ptr: felt*) {
     // Given
     alloc_locals;
     let (input) = alloc();
     let input_len = 212;
 
     let result = PrecompileBlake2f.run(0x09, input_len, input);
-    return (result.output_len, result.output);
+    memcpy(output_ptr, result.output, result.output_len);
+    return ();
 }
 
-@external
 func test_should_fail_when_flag_is_not_0_or_1{
     syscall_ptr: felt*, pedersen_ptr: HashBuiltin*, range_check_ptr, bitwise_ptr: BitwiseBuiltin*
-}() -> (output_len: felt, output: felt*) {
+}(output_ptr: felt*) {
     // Given
     alloc_locals;
     let (input) = alloc();
@@ -37,16 +32,34 @@ func test_should_fail_when_flag_is_not_0_or_1{
     assert input[212] = 0x02;
 
     let result = PrecompileBlake2f.run(0x09, input_len, input);
-    return (result.output_len, result.output);
+    memcpy(output_ptr, result.output, result.output_len);
+    return ();
 }
 
-@external
 func test_should_return_blake2f_compression{
     syscall_ptr: felt*, pedersen_ptr: HashBuiltin*, range_check_ptr, bitwise_ptr: BitwiseBuiltin*
-}(rounds: felt, h_len: felt, h: felt*, m_len: felt, m: felt*, t0: felt, t1: felt, f: felt) -> (
-    output_len: felt, output: felt*
-) {
+}(output_ptr: felt*) {
     alloc_locals;
+
+    local rounds: felt;
+    local h_len: felt;
+    let (h: felt*) = alloc();
+    local m_len: felt;
+    let (m: felt*) = alloc();
+    local t0: felt;
+    local t1: felt;
+    local f: felt;
+    %{
+        ids.rounds = program_input["rounds"]
+        ids.h_len = len(program_input["h"])
+        segments.write_arg(ids.h, program_input["h"])
+        ids.m_len = len(program_input["m"])
+        segments.write_arg(ids.m, program_input["m"])
+        ids.t0 = program_input["t0"]
+        ids.t1 = program_input["t1"]
+        ids.f = program_input["f"]
+    %}
+
     let (local input: felt*) = alloc();
     Helpers.split_word(rounds, 4, input);
     memcpy(input + 4, h, h_len);
@@ -56,6 +69,6 @@ func test_should_return_blake2f_compression{
     assert input[212] = f;
 
     let (output_len, output, gas, reverted) = PrecompileBlake2f.run(0x09, 213, input);
-
-    return (output_len=output_len, output=output);
+    memcpy(output_ptr, output, output_len);
+    return ();
 }
