+++ tests/src/kakarot/precompiles/test_blake2f.py
@@ -2,41 +2,24 @@ import random
 from typing import cast
 
 import pytest
-import pytest_asyncio
 from eth._utils.blake2.compression import TMessageBlock, blake2b_compress
-from starkware.starknet.testing.starknet import Starknet
 
 from tests.utils.helpers import pack_64_bits_little
 
 
-@pytest_asyncio.fixture(scope="module")
-async def blake2f(starknet: Starknet):
-    class_hash = await starknet.deprecated_declare(
-        source="./tests/src/kakarot/precompiles/test_blake2f.cairo",
-        cairo_path=["src"],
-        disable_hint_validation=True,
-    )
-    return await starknet.deploy(class_hash=class_hash.class_hash)
-
-
-@pytest.mark.asyncio
 @pytest.mark.BLAKE2F
 class TestBlake2f:
-    async def test_should_fail_when_input_len_is_not_213(self, blake2f):
-        (output,) = (
-            await blake2f.test_should_fail_when_input_is_not_213().call()
-        ).result
-        assert bytes(output).decode() == "Precompile: wrong input_len"
+    def test_should_fail_when_input_len_is_not_213(self, cairo_run):
+        output = cairo_run("test_should_fail_when_input_is_not_213")
+        assert bytes(output) == b"Precompile: wrong input_len"
 
-    async def test_should_fail_when_flag_is_not_0_or_1(self, blake2f):
-        (output,) = (
-            await blake2f.test_should_fail_when_flag_is_not_0_or_1().call()
-        ).result
-        assert bytes(output).decode() == "Precompile: flag error"
+    def test_should_fail_when_flag_is_not_0_or_1(self, cairo_run):
+        output = cairo_run("test_should_fail_when_flag_is_not_0_or_1")
+        assert bytes(output) == b"Precompile: flag error"
 
     @pytest.mark.parametrize("f", [0, 1])
     @pytest.mark.parametrize("seed", [0, 1, 2, 3, 4])
-    async def test_should_return_blake2f_compression(self, blake2f, f, seed):
+    def test_should_return_blake2f_compression(self, cairo_run, f, seed):
         random.seed(seed)
 
         # Given
@@ -50,13 +33,19 @@ class TestBlake2f:
         ]
 
         # When
-        got = await blake2f.test_should_return_blake2f_compression(
-            rounds, h, m, t0, t1, f
-        ).call()
+        output = cairo_run(
+            "test_should_return_blake2f_compression",
+            rounds=rounds,
+            h=h,
+            m=m,
+            t0=t0,
+            t1=t1,
+            f=f,
+        )
 
         # Then
         compress = blake2b_compress(
             rounds, cast(TMessageBlock, h_starting_state), m, (t0, t1), bool(f)
         )
         expected = [int(x) for x in compress]
-        assert got.result.output == expected
+        assert output == expected
